[RevertMe]: Workaround the composer service EACCES error
The root cause is that “Only a DRM-Master is allowed to perform mode-
setting or modify core state and only one user can be DRM-Master at a
time.”, while the early-EVS or RVC takes and holds the DRM-Master role
during system boot, so composer service cannot get it. We are discussing
the formal solution with the early-EVS/RVC owner, meanwhile proposed
this workaround solution. Btw, kernel 4.9 also used the workaround.

Change-Id: I22625fc270d46088b8e8dd1db3eafe927b21bec4
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-61242
Signed-off-by: Randy Xu <randy.xu@intel.com>
Reviewed-on: https://android.intel.com:443/632976