ASoC: SOF: Add direction param in cmd_done
There are two direction now in cmd_done, one is reply of host to DSP,
another is reply of DSP to host.The origin cmd_done is for reply of
host to DSP.

Cmd_done should be called after host has processed the reply msg
from dsp. This function is for some platforms to do extra steps to
make the following ipc msgs sent safely. Now the direction supported
by cmd_done would be host reply to dsp or dsp reply to host.

Signed-off-by: Rander Wang <rander.wang@linux.intel.com>