hyper_dmabuf/virtio: Protect virtqueue operations with spinlock.
virtqueue_add_*/virtqueue_get_buf are not safe to run in parallel as
they operate on the same structs, so code that uses them must ensure
that they won't run in parallel, as that can lead to corruption of
virtqueue data.
In case of hyper dmabuf TX queue operations can be run from different
threads and need to be protected with spinlock, RX queue operations
are always done from one thread and do not require lock.

Change-Id: I9d870bc63d202b46142b8cc0a8d86eb63cf357b8
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-63356
Signed-off-by: Mateusz Polrola <mateuszx.potrola@intel.com>
Signed-off-by: xue.yifei <yifei.xue@intel.com>
Reviewed-on: https://android.intel.com:443/642568